+++ behavior_machine/library/selector_state.py
@@ -1,12 +1,10 @@
-import threading
-from ..core import StateStatus, State, NestedState, Board
+from ..core import StateStatus, Board
 from .sequential_state import SequentialState
-import typing
 
 
 class SelectorState(SequentialState):
 
-    def execute(self, board: Board):
+    def execute(self, board: Board) -> StateStatus:
         flow_val = self.flow_in
         # execute each children one by one in order until one success.
         for child in self._children:
@@ -17,6 +15,8 @@ class SelectorState(SequentialState):
                 self._curr_child = child
                 self._curr_child.start(board, flow_val)
             self._curr_child.wait()
+            # update flow_val & set the flow out to be the current flow
+            flow_val = self._curr_child.flow_out
             # check if we are done because of interrupt
             if self.is_interrupted():
                 return StateStatus.INTERRUPTED
@@ -26,13 +26,11 @@ class SelectorState(SequentialState):
                 self.propergate_exception_information(self._curr_child)
                 return StateStatus.EXCEPTION
             elif status == StateStatus.SUCCESS:
+                self.flow_out = flow_val
                 return status
             elif status == StateStatus.FAILED:
                 pass
             else:
                 # this really shouldn't happen.
                 pass
-
-            # update flow_val
-            flow_val = self._curr_child.flow_out
         return StateStatus.FAILED